Create base class
